requirements.txt
qiskit==1.0.0
qiskit-aer==0.14.2
qiskit-ibm-runtime==0.24.0
matplotlib==3.8.0
numpy==1.24.0
python-dotenv==1.0.0
